DDF-551 Handful of small changes to allow working authentication using the new BST.
Added method to base token class to return XML version of credentials.
Added implementation-specific methods to override xml version in each subclass.
Added delegation handler for the BST tokens.
Updated the LDAP login realm to be 'ddf' rather than 'karaf'.
Updated STSRealm to be aware of the base token class - retrieving the XML version of credentials rather than blinding grabbing them - not assuming all credentials are XML.
Updated the STSCredentialsMatcher class to convert credentials to XML before comparing if necessary (rather than assuming they were always XML).
Added the BST delegation handler to the list of delegation handlers for the STS server.
Adjusted the logic for the username and password BST validator that determines whether it can handle a given token. (The realm is not being parsed from the web context as the cxf code expects, however it is being passed in the BST.