DDF-551 Replaced full SAML cookies with references to cached cookies.
Updated SAMLAuthenticationToken to hold either a full cookie or a reference.
SAMLAssertionHandler now checks for either full cookie or cookie reference and creates the corresponding SAMLAuthenticationToken.
WebSSOFilter now sets the realm value on the request as an attribute to enable easy pickin's from the LoginFilter
Added SAMLCache to cache the SAML assertions and provide lookup as references are returned in cookies.